Get WAHIS Emission Risk Factors Dataset

Helper function for getting the WAHIS emission risk factors dataset. As most analysis done require filtering for one type of each of diease, species and animal_category, this function is a helper for that.

For dataset documentation see: wahis_emission_risk_factors.

The function is supposed to be used to get data for all countries that could be involved in the introduction of a risk. The function can not be used to gather data for different diseases, species or animal categories at the same time.

Usage

get_wahis_erf(
  disease = character(),
  species = character(),
  animal_category = character()
)

Arguments

disease

filter dataset for one or more disease

species

filter dataset for one or more species

animal_category

filter dataset for one or more animal_category

Value

the emission risk factorts dataset as documented here wahis_emission_risk_factors

See also

Examples

get_wahis_erf(
  disease = "Anthrax",
  species = "Cattle",
  animal_category = "Domestic"
)
#> ✔ All data in "emission_risk_factors" valided.
#> # A tibble: 65 × 18
#>    iso3  country            disease animal_category species disease_notification
#>    <chr> <chr>              <chr>   <chr>           <chr>                  <int>
#>  1 ALB   Albania            Anthrax Domestic        Cattle                     0
#>  2 ARM   Armenia            Anthrax Domestic        Cattle                     0
#>  3 CYM   Cayman Islands     Anthrax Domestic        Cattle                     0
#>  4 HRV   Croatia            Anthrax Domestic        Cattle                     0
#>  5 FLK   Falkland Islands … Anthrax Domestic        Cattle                     0
#>  6 JPN   Japan              Anthrax Domestic        Cattle                     0
#>  7 KGZ   Kyrgyzstan         Anthrax Domestic        Cattle                     0
#>  8 MEX   Mexico             Anthrax Domestic        Cattle                     0
#>  9 RUS   Russia             Anthrax Domestic        Cattle                     0
#> 10 SYC   Seychelles         Anthrax Domestic        Cattle                     0
#> # ℹ 55 more rows
#> # ℹ 12 more variables: targeted_surveillance <int>, general_surveillance <int>,
#> #   screening <int>, precautions_at_the_borders <int>, slaughter <int>,
#> #   selective_killing_and_disposal <int>, zoning <int>,
#> #   official_vaccination <int>, last_outbreak_end_date <date>,
#> #   commerce_illegal <int>, commerce_legal <int>, data_source <chr>