Animal movement dataset mapping

A dataset mapping is used with apply_mapping() to renames, select and validation the columns in a dataset that correspond to the parameter names below.

Usage

mapping_animal_mobility(
  o_iso3 = NULL,
  o_name,
  o_lng = NULL,
  o_lat = NULL,
  d_iso3 = NULL,
  d_name,
  d_lng,
  d_lat,
  quantity = NULL
)

Arguments

o_iso3

character, optional, origin country iso3 code

o_name

character, reauired, origin name or description

o_lng

numeric, optional, origin point longitude

o_lat

numeric, option, origin point latitude

d_iso3

character, optional, destination country iso3 code

d_name

character, origin name or description

d_lng

numeric, required, destination point longitude

d_lat

numeric, required, destination point latitude

quantity

numeric, optional, used to weight animal movement flows by quantity of animals, if not provided no weighting is done.

Value

mapping object to be used with apply_mapping()

See also