Emission risk factors dataset mapping

A dataset mapping is used with apply_mapping() to renames, select and validation the columns in a dataset that correspond to the parameter names below.

Usage

mapping_emission_risk_factors(
  iso3,
  country,
  disease,
  animal_category,
  species,
  disease_notification,
  targeted_surveillance,
  general_surveillance,
  screening,
  precautions_at_the_borders,
  slaughter,
  selective_killing_and_disposal,
  zoning,
  official_vaccination,
  last_outbreak_end_date,
  commerce_illegal,
  commerce_legal,
  data_source = NULL
)

Arguments

iso3

character, required, country iso3 code

country

character, required, country name

disease

character, required, disease name

animal_category

character, required, animal category "Wild" or "Domestic"

species

character, required, animal species name

disease_notification

integer, required, must be NA, 1 or 0: 1 representing this policy is not enforced in that country (i.e. 1 is risky, 0 is not).

targeted_surveillance

as disease_notification

general_surveillance

as disease_notification

screening

as disease_notification

precautions_at_the_borders

as disease_notification

slaughter

as disease_notification

selective_killing_and_disposal

as disease_notification

zoning

as disease_notification

official_vaccination

as disease_notification

last_outbreak_end_date

date, required, must be a date: represents the last time there was an outbreak of this disease in this country.

commerce_illegal

integer, required, must be NA, 1 or 0: 1 representing this type of commerce is conducted between this country and country for which risk of introduction is been calcalated (i.e. 1 is risky, 0 is not).

commerce_legal

as commerce_illegal

data_source

character, optional, used to keep track of where the data is sourced.

Value

mapping object to be used with apply_mapping()

Details

For more about the data requirements of this dataset see wahis_emission_risk_factors documentation