Read Emission Risk Factor File

Reads a text file containing emission risk factors for animal diseases. The file should be in a tabular format with specific columns (see details).

Usage

read_emission_risk_factor_file(filepath)

Arguments

filepath

A character string specifying the path to the text file.

Details

The text file is expected to be in a tabular format with the following columns:

Country information:

iso3: character - A three-letter ISO 3166-1 alpha-3 country code.

country: character - The full name of the country.

Disease information (should be the same for the whole dataset):

disease: character - The name of the disease.

animal_category: character - The category of animal ("Domestic" or "Wild").

species: character - The species name affected by disease.

Measures of control and surveillence, all are either 1 or 0 (0 = measure inplace, 1 = no measure inplace, i.e. 1 = there is a risk, 0 = there is no risk).

disease_notification: integer - Indicator of notification

targeted_surveillance: integer - Risk factor score for targeted surveillance efforts.

general_surveillance: integer - Risk factor score for general surveillance activities.

screening: integer - Risk factor score for screening measures.

precautions_at_the_borders: integer - Risk factor score for precautions taken at borders.

slaughter: integer - Risk factor score related to the slaughter process.

selective_killing_and_disposal: integer - Risk factor score for selective killing and disposal procedures.

zoning: integer - Risk factor score for zoning strategies.

official_vaccination: integer - Risk factor score for official vaccination programs.

Epidemiological status:

last_outbreak_end_date: Date (YYYY-MM-DD) - The end date of the last outbreak.

Animal commerce (indicators of level of illegality of commerce)

commerce_illegal: integer - either 1 or 0

commerce_legal: integer - either 1 or 0

Other information:

data_source: character - Reference or source of the data.

Examples

tun_erf_file <-
  system.file(
    package = "riskintrodata",
    "samples",
    "tunisia",
    "emission_risk_factor",
    "emission_risk_factors.csv"
  )
x <- read_emission_risk_factor_file(tun_erf_file)
x
#> # A tibble: 65 × 18
#>    iso3  country            disease animal_category species disease_notification
#>    <chr> <chr>              <chr>   <chr>           <chr>                  <int>
#>  1 ALB   Albania            Brucel… Domestic        Cattle                     0
#>  2 BWA   Botswana           Brucel… Domestic        Cattle                     1
#>  3 CYM   Cayman Islands     Brucel… Domestic        Cattle                     0
#>  4 JPN   Japan              Brucel… Domestic        Cattle                     0
#>  5 SYC   Seychelles         Brucel… Domestic        Cattle                     0
#>  6 SYR   Syria              Brucel… Domestic        Cattle                     0
#>  7 PSE   Palestine          Brucel… Domestic        Cattle                     0
#>  8 USA   United States of … Brucel… Domestic        Cattle                     0
#>  9 ARM   Armenia            Brucel… Domestic        Cattle                     0
#> 10 AZE   Azerbaijan         Brucel… Domestic        Cattle                     0
#> # ℹ 55 more rows
#> # ℹ 12 more variables: targeted_surveillance <int>, general_surveillance <int>,
#> #   screening <int>, precautions_at_the_borders <int>, slaughter <int>,
#> #   selective_killing_and_disposal <int>, zoning <int>,
#> #   official_vaccination <int>, last_outbreak_end_date <date>,
#> #   commerce_illegal <int>, commerce_legal <int>, data_source <chr>